+++ lib/ask-tokens-rails.rb
@@ -2,4 +2,5 @@
 
 require "ask-tokens"
 require "ask/tokens/rails/version"
+require "ask/tokens/rails/configuration"
 require "ask/tokens/rails/railtie"