+++ lib/generators/ask_tokens/install/install_generator.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "ask-tokens-rails"
 require "rails/generators"
 require "rails/generators/migration"